Make dockerfile for building frontend

+++ database/migrations/2024_09_20_064358_create_projects_table.php
@@ -30,8 +30,8 @@ return new class extends Migration {
 
 
         Schema::create('project_user', function (Blueprint $table) {
-            $table->unsignedInteger('project_id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('project_id');
+            $table->unsignedBigInteger('user_id');
             $table->timestamps();
             $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
